made Twiddle::invSqrtMagic explicit cast for pedantic compilers

+++ Gamma/scl.h
@@ -61,8 +61,8 @@ template<int N, class T> struct InvSqrtNewton
 };
 
 template<class T> const Twiddle<T> invSqrtMagic();
-template<> inline const Twiddle<float > invSqrtMagic(){ return Twiddle<float >(0x5f3759df); }
-template<> inline const Twiddle<double> invSqrtMagic(){ return Twiddle<double>(0x5fe6ec85e7de30daULL); }
+template<> inline const Twiddle<float > invSqrtMagic(){ return Twiddle<float >((float)0x5f3759df); }
+template<> inline const Twiddle<double> invSqrtMagic(){ return Twiddle<double>((double)0x5fe6ec85e7de30daULL); }
 
 
 /// Approximate square root using a quick log base-2 method.
